requirements.txt
psycopg2 >= 2.9.9
pyTelegramBotAPI >= 4.16.1
alembic >= 1.13.1
pyyaml >= 6.0.1
sqlalchemy >= 2.0.29
python-json-logger >= 2.0.7
asyncpg >= 0.29.0